requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2
python-dotenv>=1.0
gunicorn
fpdf2
